Fixed 100+ warnings.
Mostly float/integral number conversions and incorect float literal.
Some signed/unsigned conversion fixes (loop "optimizations").

+++ YetAnotherPlayground/src/Camera.h
@@ -76,7 +76,7 @@ private:
 	glm::vec2 viewportSize;
 
 public:
-	void windowDidResize(float width, float height);
+	void windowDidResize(int width, int height);
 	void setupViewport(float x, float y, float widthm, float height);
 	void loadViewport();
